move api folder to higher level to allow cross feature sharing
