+++ apps/web/src/features/search/types/search.types.ts
@@ -5,27 +5,6 @@ export type SearchCriteria = {
   keyword: string;
 };
 
-export type JobCardDto = {
-  id: string | number;
-  title: string;
-  company: string;
-  locationName: string;
-  summary: string;
-  url: string;
-  sourceName: string;
-  isRemote: boolean;
-  postedAtUtc: string;
-};
-
-export type SearchJobsResponseDto = {
-  postcode: string;
-  keyword: string;
-  pageIndex: number;
-  pageSize: number;
-  totalCount: number;
-  jobs: JobCardDto[];
-};
-
 export type SearchViewModel = {
   postcode: string;
   keyword: string;
